The strongbox can be opened and the bag can be picked
Updated the script with the logic for the Basement so after a combination is done a wait is triggered to give feedback to the player.
When all combinations are done the strongbox opens and the bag appears.
The bag can be added to the inventory by clicking the strongbox once it is open.
The number of combinations required to open the strongbox can be set in the Editor using a range field.