Project hierarchy updated.
The new hierarchy removes containers that won't be used by a 2D game and put all the elements related to an object as childs of it (i.e. a hotspot has its markers and interactions as childs of itself).
Removed GitHub plugin until it stabilizes.
All texts translated to spanish. First spanish, then other languages.